RAG/pyproject.toml
[project]
name = "langgraph-rag-agent"
version = "0.1.0"
description = "LangGraph-based RAG Agent with explicit routing between RAG and Analysis paths"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.0",
    "typing-extensions>=4.5.0",
    "tiktoken>=0.5.0",
    "python-docx>=1.0.0",
    "pyyaml>=6.0",
    "numpy>=1.21.0",
    "zhipuai>=2.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "sniffio>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
future = [
    "dspy-ai",  # For future DSPy integration
    "langchain>=0.1.0",  # For future LLM integration
    "langchain-openai",  # For OpenAI models
    "faiss-cpu",  # For vector database
    "chromadb",  # Alternative vector database
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.uv]
package = false
